my courses

+++ Server/src/main/java/bgu/spl/net/impl/BGRSServer/User.java
@@ -45,7 +45,7 @@ public class User {
     }
 
     public void unregisterCourse(int courseNum){
-        userCourses.remove(courseNum);
+        userCourses.remove((Integer) courseNum);
     }
 
     /**
